requirements.txt
altgraph==0.17.4
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dnspython==2.6.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
lxml==5.2.1
oauthlib==3.2.2
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pymongo==4.7.2
pymupdf==1.24.3
pymupdfb==1.24.3
pyparsing==3.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.1
xlsxwriter==3.2.0
